+++ tests/Public/Get-ShlinkDomains.Acceptance.Tests.ps1
@@ -1,0 +1,24 @@
+BeforeAll {
+    # Always use built code if running in a pipeline
+    if ($env:USER -eq 'runner') {
+        Import-Module "$PSScriptRoot/../../build/PSShlink/PSShlink.psd1" -Force
+    }
+    # Check if module is already imported, as it can be via VSCode task where you can choose what code base to test
+    # and you might not want to cloober it with the non-built code
+    elseif (-not (Get-Module PSShlink)) {
+        Import-Module "$PSScriptRoot/../../src/PSShlink.psd1" -Force
+    }
+}
+
+Describe "Get-ShlinkDomains" {
+    It "Returns domains configured with Shlink" {
+        $Params = @{
+            ShlinkServer = $env:ShlinkServer
+            ShlinkApiKey = $env:ShlinkAPIKey | ConvertTo-SecureString
+            ErrorAction  = 'Stop'
+        }
+        $Object = Get-ShlinkDomains @Params
+        $Object.data.domain    | Should -Be 'psshlink.codaamok'
+        $Object.data.isDefault | Should -BeTrue
+    }
+}